refactor(mode): ♻️ restructure model in clearer directories
Continues #726
