+++ lib/music_notes.dart
@@ -30,7 +30,10 @@ export 'src/key/german_key_notation.dart';
 export 'src/key/key.dart';
 export 'src/key/romance_key_notation.dart';
 export 'src/key_signature/key_signature.dart';
+export 'src/mode/english_tonal_mode_notation.dart';
+export 'src/mode/german_tonal_mode_notation.dart';
 export 'src/mode/mode.dart';
+export 'src/mode/romance_tonal_mode_notation.dart';
 export 'src/notation_system/notation_system.dart';
 export 'src/note/english_note_notation.dart';
 export 'src/note/german_note_notation.dart';
